requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
colorama>=0.4.6
tqdm>=4.66.0
Flask>=3.0.0